OAK-10973 - Performance tune property compression/decompression set decompressedValue
