+++ oak-store-document/src/main/java/org/apache/jackrabbit/oak/plugins/document/CompressedDocumentPropertyState.java
@@ -19,6 +19,12 @@ package org.apache.jackrabbit.oak.plugins.document;
 import static java.util.Collections.emptyList;
 import static org.apache.jackrabbit.oak.plugins.memory.PropertyStates.createProperty;
 
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.io.OutputStream;
+import java.nio.charset.StandardCharsets;
+import java.util.Arrays;
 import java.util.List;
 
 import javax.jcr.PropertyType;
@@ -26,8 +32,11 @@ import javax.jcr.PropertyType;
 import org.apache.jackrabbit.guava.common.collect.Lists;
 import org.apache.jackrabbit.oak.api.PropertyState;
 import org.apache.jackrabbit.oak.api.Type;
+import org.apache.jackrabbit.oak.commons.Compression;
+import org.apache.jackrabbit.oak.commons.LongUtils;
 import org.apache.jackrabbit.oak.commons.json.JsopReader;
 import org.apache.jackrabbit.oak.commons.json.JsopTokenizer;
+import org.apache.jackrabbit.oak.commons.properties.SystemPropertySupplier;
 import org.apache.jackrabbit.oak.json.TypeCodes;
 import org.apache.jackrabbit.oak.plugins.memory.AbstractPropertyState;
 import org.apache.jackrabbit.oak.plugins.memory.BinaryPropertyState;
@@ -38,11 +47,15 @@ import org.apache.jackrabbit.oak.plugins.memory.PropertyStates;
 import org.apache.jackrabbit.oak.plugins.memory.StringPropertyState;
 import org.apache.jackrabbit.oak.plugins.value.Conversions;
 import org.jetbrains.annotations.NotNull;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * PropertyState implementation with lazy parsing of the JSOP encoded value.
  */
-final class DocumentPropertyState implements PropertyState {
+public final class CompressedDocumentPropertyState implements PropertyState {
+
+    private static final Logger LOG = LoggerFactory.getLogger(CompressedDocumentPropertyState.class);
 
     private final DocumentNodeStore store;
 
@@ -51,11 +64,37 @@ final class DocumentPropertyState implements PropertyState {
     private final String value;
 
     private PropertyState parsed;
+    private final byte[] compressedValue;
+    private final Compression compression;
+    private String decompressedValue;
 
-    DocumentPropertyState(DocumentNodeStore store, String name, String value) {
+    private static int COMPRESSION_THRESHOLD = SystemPropertySupplier
+            .create("oak.documentMK.stringCompressionThreshold ", -1).loggingTo(LOG).get();
+
+    CompressedDocumentPropertyState(DocumentNodeStore store, String name, String value, Compression compression) {
         this.store = store;
         this.name = name;
+        this.compression = compression;
+        int size = value.length();
+        byte[] compressedValue = null;
+        if (size > COMPRESSION_THRESHOLD) {
+            try {
+                compressedValue = compress(value.getBytes(StandardCharsets.UTF_8));
+                value = null;
+            } catch (IOException e) {
+                LOG.warn("Failed to compress property {} value: ", name, e);
+            }
+        }
         this.value = value;
+        this.compressedValue = compressedValue;
+    }
+
+    private byte[] compress(byte[] value) throws IOException {
+        ByteArrayOutputStream out = new ByteArrayOutputStream();
+        OutputStream compressionOutputStream = compression.getOutputStream(out);
+        compressionOutputStream.write(value);
+        compressionOutputStream.close();
+        return out.toByteArray();
     }
 
     @NotNull
@@ -116,7 +155,23 @@ final class DocumentPropertyState implements PropertyState {
      */
     @NotNull
     String getValue() {
-        return value;
+        return value != null ? value : decompress(this.compressedValue);
+    }
+
+    private String decompress(byte[] value) {
+        if (decompressedValue == null) {
+            try {
+                decompressedValue = new String(compression.getInputStream(new ByteArrayInputStream(value)).readAllBytes(), StandardCharsets.UTF_8);
+            } catch (IOException e) {
+                LOG.error("Failed to decompress property {} value: ", getName(), e);
+                decompressedValue = "\"{}\"";
+            }
+        }
+        return decompressedValue;
+    }
+
+    byte[] getCompressedValue() {
+        return compressedValue;
     }
 
     //------------------------------------------------------------< Object >--
@@ -125,10 +180,20 @@ final class DocumentPropertyState implements PropertyState {
     public boolean equals(Object object) {
         if (this == object) {
             return true;
-        } else if (object instanceof DocumentPropertyState) {
-            DocumentPropertyState other = (DocumentPropertyState) object;
-            return this.name.equals(other.name)
-                    && this.value.equals(other.value);
+        } else if (object instanceof CompressedDocumentPropertyState) {
+            CompressedDocumentPropertyState other = (CompressedDocumentPropertyState) object;
+            if (!this.name.equals(other.name) || !Arrays.equals(this.compressedValue, other.compressedValue)) {
+                return false;
+            }
+            if (this.compressedValue == null && other.compressedValue == null) {
+                return value.equals(other.value);
+            } else {
+                // Compare length and content of compressed values
+                if (this.compressedValue.length != other.compressedValue.length) {
+                    return false;
+                }
+                return Arrays.equals(this.compressedValue, other.compressedValue);
+            }
         }
         // fall back to default equality check in AbstractPropertyState
         return object instanceof PropertyState
@@ -145,11 +210,15 @@ final class DocumentPropertyState implements PropertyState {
         return AbstractPropertyState.toString(this);
     }
 
+    static void setCompressionThreshold(int compressionThreshold) {
+        COMPRESSION_THRESHOLD = compressionThreshold;
+    }
+
     //----------------------------< internal >----------------------------------
 
     private PropertyState parsed() {
         if (parsed == null) {
-            JsopReader reader = new JsopTokenizer(value);
+            JsopReader reader = new JsopTokenizer(getValue());
             if (reader.matches('[')) {
                 parsed = readArrayProperty(name, reader);
             } else {
@@ -180,10 +249,11 @@ final class DocumentPropertyState implements PropertyState {
     static PropertyState readProperty(String name, DocumentNodeStore store, JsopReader reader) {
         if (reader.matches(JsopReader.NUMBER)) {
             String number = reader.getToken();
-            try {
-                return new LongPropertyState(name, Long.parseLong(number));
-            } catch (NumberFormatException e) {
+            Long maybeLong = LongUtils.tryParse(number);
+            if (maybeLong == null) {
                 return new DoublePropertyState(name, Double.parseDouble(number));
+            } else {
+                return new LongPropertyState(name, maybeLong);
             }
         } else if (reader.matches(JsopReader.TRUE)) {
             return BooleanPropertyState.booleanProperty(name, true);
@@ -238,12 +308,13 @@ final class DocumentPropertyState implements PropertyState {
         while (!reader.matches(']')) {
             if (reader.matches(JsopReader.NUMBER)) {
                 String number = reader.getToken();
-                try {
-                    type = PropertyType.LONG;
-                    values.add(Long.parseLong(number));
-                } catch (NumberFormatException e) {
+                Long maybeLong = LongUtils.tryParse(number);
+                if (maybeLong == null) {
                     type = PropertyType.DOUBLE;
                     values.add(Double.parseDouble(number));
+                } else {
+                    type = PropertyType.LONG;
+                    values.add(maybeLong);
                 }
             } else if (reader.matches(JsopReader.TRUE)) {
                 type = PropertyType.BOOLEAN;
@@ -278,3 +349,10 @@ final class DocumentPropertyState implements PropertyState {
         return createProperty(name, values, Type.fromTag(type, true));
     }
 }
+
+
+
+
+
+
+
